separate config and context files
